backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gspread==6.1.4
google-auth==2.29.0
gunicorn==22.0.0
python-dotenv==1.0.1
